server/ecommerce/requirements.txt
djangorestframework
gunicorn
whitenoise